Fix ORCA SubLink processing at query normalization
The ORCA optimizer could produce an invalid query plan or fallback to the legacy
optimizer due to incorrect processing of nested SubLinks (SubLink contains one
more SubLink or rtable subquery inside) during query normalization.

Firstly, the incorrect behaviour took place when a query had a nested
SubLink in its targetList, and SubLink lower levels contained attributes that
were referenced in the query's GROUP BY clause. ORCA could construct for
such queries an invalid plan, according to which some wrong and unexpected
attributes got into the nested part of SubLink's plan (the example is provided
in tests section). And the presence of those attributes could lead to incorrect
query execution or even segmentation fault.

Inside the NormalizeGroupByProjList function the context.m_lower_table_tlist,
which contains grouping attributes from target_list_copy, is created. And the
entries of target_list_copy, that are not grouping attributes, are mutated
inside the RunExtractAggregatesMutator in order to make their inner expressions
reference the corresponding entries from context.m_lower_table_tlist.
The invalid plan could be produced because of incorrect mapping between Var's
varattno (in target_list_copy) and some TargetEntry's resno
(in context.m_lower_table_tlist). This mapping is established by
RunExtractAggregatesMutator, where for the Var code branch the following
assignment takes place: var->varattno = found_tle->resno, where found_tle is the
corresponding entry from context.m_lower_table_tlist. However, the behaviour of
this function was erroneous for the case of nested SubLinks because it
mutated only the upper level of nested SubLink structure, and other subqueries
remained unmodified because RunExtractAggregatesMutator did not have a branch
for the Query, which was required for futher diving into lower subqueries. And
because inner Vars hadn't been mutated, they started reference wrong attributes.

This patch modifies RunExtractAggregatesMutator function by adding the Query
branch to make the mutator correctly step into nested SubLink queries. And the
call of MutateQueryOrExpressionTree is replaced to the call of mutator itself
in order to correctly modify m_current_query_level when stepping into lower
subqueries.

Secondly, the erroneous behaviour took place when nested SubLink was in
GROUP BY clause itself (see the test section). In this case nested SubLinks were
mutated incorrectly by RunGroupingColMutator.
RunGroupingColMutator had Query branch, however, when mutating the subquery
of the SubLink (at second query level), the m_current_query_level variable was
not increased respectively, because MutateQueryOrExpressionTree, which was
responsible for calling the RunGroupingColMutator on the lower subquery, did not
know anything about a context. Therefore, for the query like in the tests, wrong
value of m_current_query_level led to unnecessary modification of the Var's
varlevelsup value. And when Var's varlevelsup value had been unnecessary
modified, it started referencing different query level. This could lead to the
fallback or invalid plan construction at further planning stages.

This patch fixes erroneous behaviour of RunGroupingColMutator by replacing the
call of MutateQueryOrExpressionTree to the call of RunGroupingColMutator in
order to properly modify m_current_query_level when mutating lower level rtes.

Thirdly, similar issue could arise when query had correlated subquery in SubLink
under the AggRef.
While mutating the SubLink inside RunExtractAggregatesMutator, the Aggref branch
is executed for SubLink's query targetList, which mutates args of the AggRef.
And if a Var had varlevelsup value greater than the agglevelsup, i.e the Var
referenced a relation that is higher than the AggRef, the Var was not mutated
at all. The varlevelsup field was not modified and because of that, when AggRef
was pulled up to the zero query level (by appending it to the
context.m_lower_table_tlist), the Var, whose varlevelsup was higher than
AggRef's, and which remained unchanged, started referencing the wrong relation.
This could lead to the fallback or invalid plan construction when processing
attributes at further planning stages.

This patch makes the RunExtractAggregatesMutator modify the varlevelsup of the
Var relatively the AggRef level in order to preserve proper query level.