Make InitPlans non-rescannable
The executor enabled the EXEC_FLAG_REWIND flag for all types of SubPlans: either
for InitPlans or correlated/uncorrelated SubPlans. This flag represents that the
rescan is expected and is used during initialization of executor state. However,
if a query had InitPlans, which contained non-rescannable nodes (like Split
Update in the tests), the executor failed with assertion error (for example,
when calling ExecInitSplitUpdate when initializing executor state inside the
ExecInitNode for the InitPlan).

Because InitPlans are essentially executed only once, there is no need to expect
a rescan of the InitPlan. Therefore, in order to support non-rescannable
operations inside the InitPlans, this patch disables EXEC_FLAG_REWIND flag for
InitPlans. Before the ExecInitNode is called for a SubPlan, the if-clause, that
checks whether this SubPlan is an InitPlan, is added. This condition checks the
occurence of subplan_idx inside the pre collected set of InitPlan plan_id. This
set is formed by executing additional walker getInitPlans, that collects all
InitPlan ids from the plan tree.