Avoid rescan of modifying operations inside correlated Subplans.
When a query had a modifying command inside the correlated SubPlan, the
ModifyTable node could be rescanned for each outer tuple. That lead to
execution errors (rescan of specific nodes is not supported). This happened
because the ParallelizeCorrelatedSubplanMutator function did not expect the
ModifyTable node inside the correlated SubPlans.

This patch adds the handling of ModifyTable node for correlated SubPlans. The
ParallelizeCorrelatedSubplanMutator function is extended with the ModifyTable
code branch, that simply broadcasts or focuses the result of ModifyTable,
depending on the target flow type. Then the result is materialized in order
to avoid rescan of underlying nodes.