Support Explicit Redistribute Motion inside the SubPlans.
When a query had UPDATE/DELETE operation, that required Explicit Redistribute,
inside the SubPlan (or InitPlan), the Ecxplicit Redistribute did not applied
due to an assertion error. The error was thrown from apply_motion_mutator
function and said that the Flow's segidColIdx attribute did not satisfy a
validity check. That happened because, previously SubPlans had been mutated
inside the ParallelizeSubplan function, where the SubPlan's plan had been
copied via copyObject function. This function copies the whole plan including
the Flow node. However the Flow node copying did not include the copying of
segidColIdx Flow field, which is used for correct performing of Explicit
Redistribute Motion.

This patch solves the issue by adding segidColIdx to the list of attributes in
the copyFlow function.