Support Explicit Redistribute Motion inside the SubPlans.
When a query had UPDATE/DELETE operation, that required Explicit Redistribute,
inside the SubPlan (or InitPlan), the Explicit Redistribute did not applied
correctly. On debug build with asserts the assertion error was thrown from
apply_motion_mutator function, which said that the Flow's segidColIdx attribute
did not point to any element of plan's (on top of which Explicit Redistribute
was meant to be applied) target list. This field represents an
index of the gp_segment_id column in the target list, and this column is
subsequently used for Explicit Redistribute Motion. On release build without
asserts executor could not execute the Explicit Redistribute Motion by
throwing an error related to invalid attribute number or could perform Explicit
Redistribute Motion wrongly because of segidColIdx referencing a wrong column.

The problem occured because previously SubPlans had been mutated inside the
ParallelizeSubplan function, where the SubPlan's plan had been copied via
copyObject function. This function copies the whole plan including the Flow
node. However the Flow node copying did not include the copying of segidColIdx
Flow field, which is used for valid performance of Explicit Redistribute Motion.

This patch solves the issue by adding segidColIdx to the list of attributes in
the copyFlow function.