Fix Flow structure copying, comparison, serialization and deserialization.
Flow structure has a segidColIdx field, which referencing a gp_segment_id
column in plan's targetList when the Explicit Redistribute Motion is requested.
There was a problem, that _copyFlow, _equalFlow, _outFlow and _readFlow
functions did not handled the segidColIdx field. Therefore, the Flow node
could not be serialized and deserialized, or copied corretly.

The problem manifested itself when a query had UPDATE/DELETE operation, that
required Explicit Redistribute, inside the SubPlan (or InitPlan). The Explicit
Redistribute did not applied correctly inside the apply_motion_mutator function
because by that moment the value segidColIdx had been lost. The problem occured
because previously SubPlans had been mutated inside the ParallelizeSubplan
function, where the SubPlan's plan had been copied via copyObject function.
This function copies the whole plan including the Flow node, which is copied
via _copyFlow function. However the Flow node copying did not include the
copying of segidColIdx Flow field, which is used for valid performance of
Explicit Redistribute Motion.

Therefore, this patch solves the issue by adding segidColIdx to the list of
fields to copy, serialize, deserialize and compare in the _copyFlow, _outFlow,
_readFlow and _equalFlow functions respectively.