Make InitPlans non-rescannable
The executor enabled the EXEC_FLAG_REWIND flag for all types of SubPlans: either
for InitPlans or correlated/uncorrelated SubPlans. This flag represents that the
rescan is expected and is used during initialization of executor state. However,
if a query had InitPlans, which contained non-rescannable nodes (like Split
Update in the tests), the executor failed with assertion error (for example,
when calling ExecInitSplitUpdate when initializing executor state inside the
ExecInitNode for the InitPlan).

Because InitPlans are essentially executed only once, there is no need to expect
a rescan of the InitPlan. Therefore, in order to support non-rescannable
operations inside the InitPlans, this patch disables EXEC_FLAG_REWIND flag for
InitPlans.

This patch partially returns vanilla postgres logic, which used
plannedstmt->rewindPlanIDs bitmapset for making a decision whether current
SubPlan should be executed with EXEC_REWIND flag. This bitmapset used to be
filled with the ids of such SubPlans, that could optimize the rescan
operation if the EXEC_REWIND is set, like parameterless subplans. Other
types of SubPlans were considered rescannable by default and there were
no need to set the EXEC_REWIND flag for them. However, GPDB interprets the
EXEC_REWIND flag as an indicator that the node is likely to be rescanned, and
also used this flag to delay the eager free. Therefore, this patch proposes
to fill plannedstmt->rewindPlanIDs set with all the subplans ids, except
InitPlans, and to set EXEC_REWIND flag only for those subplans that are
in the rewindPlanIDs bitmapset.

As for legacy optimizer, the if-clause influencing the filling of the
bitmapset is changed inside the build_subplan function in order to filter out
any InitPlans.

As for ORCA optimizer, rewindPlanIDs was not previously used, and this patch
adds a bunch of logic to fill this bitmapset with subplan ids. Firstly, the
CContextDXLToPlStmt class is extended with an extra member m_subplan_exprs_list
that consists of expressions ot SubPlan type. This list is filled via
AddSubPlanExpr member function, which is called from the
TranslateSubplanFromChildPlan function. Also the getter for the
m_subplan_exprs_list is added. Next, the CTranslatorDXLToPlStmt is
extended with SetRewindPlanIds function, that performs the filtration
of the m_subplan_exprs_list and fills the planned_stmt->rewindPlanIDs
with subplan ids, that correspond to any SubPlan, except InitPlans.
The SetRewindPlanIds function is called at one of the final planning
stages, where DXL plan tree in converted to the PlannedStmt. Thus,
proposed changes allows to correctly fill the planned_stmt->rewindPlanIDs
bitmapset.

At executor side, the condition checking whether the SubPlan is
in the planned_stmt->rewindPlanIDs is added to the function InitPlan.
From that point, SubPlans will be initialized with EXEC_REWIND flag
only if they are not InitPlans.