Make InitPlans non-rescannable
The executor enabled the EXEC_FLAG_REWIND flag for all types of SubPlans: either
for InitPlans or correlated/uncorrelated SubPlans. This flag represents that the
rescan is expected and is used during initialization of executor state. However,
if a query had InitPlans, which contained non-rescannable nodes (like Split
Update in the tests), the executor failed with assertion error (for example,
when calling ExecInitSplitUpdate when initializing executor state inside the
ExecInitNode for the InitPlan).

Because InitPlans are essentially executed only once, there is no need to expect
a rescan of the InitPlan. Therefore, in order to support non-rescannable
operations inside the InitPlans, this patch disables EXEC_FLAG_REWIND flag for
InitPlans.

This patch turns back to vanilla postgres field of PlannedStmt structure,
plannedstmt->rewindPlanIDs bitmapset, that was used for making a decision
whether current SubPlan should be executed with EXEC_REWIND flag. This bitmapset
used to be filled with the ids of such SubPlans, that could optimize the rescan
operation if the EXEC_REWIND is set, like parameterless subplans. Other
types of SubPlans were considered rescannable by default and there were
no need to set the EXEC_REWIND flag for them. However, GPDB interprets the
EXEC_REWIND flag as an indicator that the node is likely to be rescanned, and
also uses this flag to delay the eager free. Therefore, this patch proposes
to fill plannedstmt->rewindPlanIDs set with all the subplans ids, except
InitPlans, and to set EXEC_REWIND flag only for those subplans that are
in the rewindPlanIDs bitmapset.

For legacy optimizer, the if-clause influencing the filling of the
bitmapset is changed inside the build_subplan function in order to filter out
any InitPlans.
As for ORCA optimizer, rewindPlanIDs was not previously used, and, moreover,
ORCA is unable to produce InitPlans. Therefore, all SubPlans for plans
generated by the optimizer must be initialized with EXEC_REWIND flag.

At executor side, the condition checking whether the SubPlan is
in the planned_stmt->rewindPlanIDs is added to the InitPlan function.
For ORCA generated plans the filtration is not held, and EXEC_REWIND
is always set as it was before.
From that point, SubPlans will be initialized with EXEC_REWIND flag
only if they are not InitPlans.