fixed the that ORCA did not check the availability of the index in the current transaction as the postgres optimizer does
When updating a row in heap tables, a HOT (Heap-Only Tuple) chain is created
where the old version of the row points to the new version. When creating an
index on such a table and there is already another index, the new index's
elements will point to the beginning of the chain (the old value), while having
the new value. In a situation where we have a parallel transaction reading this
table using the new index, we may find the old row. for more info, see
src/backend/access/heap/README.HOT
example:
In the initial state, we have a tuple that the index points to.

In another session, we open a transaction with a repeatable read isolation level
and do a select.

in session 1, we update the tuple, thereby creating a HOT chain and creating an
index for the updated column. the new index will reference the first version of
the tuple.

if now in session 2 we do a search for a new index with a new value, then we
will find the old tuple, since for this session the first version is alive.

The PostgreSQL optimizer handles this situation correctly and ignores the new
index in the old transaction. ORCA does not.

This patch fixes this by adding a visibility check for the index in ORCA. Also,
in such a case, we mark the cache and the plan as transient to reset the mdcache
in ORCA and the cached plan in the case of a prepare at the beginning of the
next transaction.