Fix deadlock by shared SegmentGeneral CTE
Depending on the usage of the shared CTE with SegmentGeneral subplan, the
number of segments the CTE are executed on may be single for joins with
Singleton nodes, and multiple for hashed nodes. In the current implementation
this may lead to deadlock if the CTE is used for both join targets: the Join
with the Singleton node results in the Share Input Scan producer being executed
on a single segment, while the Join with the Hashed node creates
Share Input Scan reader on multiple segments, so the plan execution hangs.
If we force execution of CTE on multiple segments as well, it will cause
redundant motions in case of joining the CTE with another SegmentGeneral.
At the moment of constructing and sharing the CTE we don't know the rest of the
plan, so we can't predict the correct CTE locus. Because replicated tables are
considered small, the most universal and optimal way to fix deadlock would be
to inline CTE scans with SegmentGeneral locus.

This patch fixes the deadlock by disabling sharing CTE if the subplan has
SegmentGeneral locus.