Implement vacuum/analyze ordering functions for arenadata_toolkit
This patch extends the functionality of the arenadata_toolkit extension
by adding two functions to obtain different table orders for vacuum/analyze.
These two functions are planned to be used in the vacuum command in
the operation.py script. By default, the bundle will use the current approach,
in which new tables are placed at the beginning of the list. But users will
be able to reconfigure it to use a second function that places new tables
at the end of the list, or even use a custom query to determine the order.