Forbidden volatile functions over Replicated locus for joins
This commit introduces the variant of volatile functions handling,
when Replicated locus is joined with others. In order to avoid the
Explicit Gather - Broadcast/Redistribute pair, where 2 motions in
the plan are one on the top of the other, we decide to throw an
error in cases, when any of join operands has Replicated locus
and the path contains volatile functions inside.
path_contains_volatile_functions function was added to pathnode.c,
this function tries to cover vast variety of queries and checks
whether there are volatile functions in the path's restrictinfo,join
restrictinfo, targetlist and qual of subqueries' subplans.