Fixed docker files for running the gpdb7 test suite
Dnf is now used instead of yum as a package manager.
Devtoolset-7 was removed since the necessary tools are available in the main
repository.
Added installation of llvm and clang for building with JIT support.
RSA1 key generation was removed since it's deprecated.
The mdw container was renamed to cdw.
Now, instead of a combination of bash and tail, the init process and sleep
infinity are used in containers to prevent zombie processes from appearing.