Truncate the first segment file for temporary relations
Problem description:
After the drop of the temporary heap relation, disk space was not released to
the OS for several minutes. The expected behavior is immediate disk space
release.

Root cause:
Background writer process kept file descriptors of the relation open until
the next checkpoint. The next checkpoint occurred after a timeout, which could
be up to CheckPointTimeout seconds (300 by default).
Only temporary relations suffered from it because, in mdunlinkfork first
segment file was truncated for all non-temporary relations.

Fix:
In mdunlinkfork, truncate the first segment file for all relations, including
temporary. A test case was added to the tablespace test to catch such a
situation. Plus, some code refactoring in the tablespace test was done.