Check for Motions before scan if requesting Explicit Redistribute Motion
Previously, `Explicit Redistribute Motion` was added if there were any motions
in the subplan, ignoring any motions in `InitPlans`.

This patch updates the logic: if an `Explicit Redistribute Motion` has no
motions underneath it's subtree, then the row to update must originate from the
same segment, and no `Explicit Redistribute Motion` is needed. If there are any
motions, `Explicit Redistribute Motion` should be added only if there is a
motion between the scan and the `ModifyTable` on the relation we are going to
update.

Even if previous Motions were performed on a leaf partition or inherited table,
targetlists from Motions refer to relids of their parents. So, if relid has a
superclass, we should check for inheritance instead of comparing Oids.