Look into hashExprs for Redistribute Motion
While checking the assumption that the right subtree is more frequent, I found
out that target list of `Redistribute Motion` may contain columns which are not
used for distribution. Behavior was correct only because the right subtree is
checked first. In cases when the scan would be in the left subtree, this is
incorrect, and `hashExprs` need to be checked instead.

Checking `hashExprs` may contain different structures that inherit `Expr` node.
The structure we need is of type `Var`, because it's the only child of `Expr`
node that we can get relation ID from.