Move for_each_cell under the null check
This block of code is executed when ALTER TABLE ADD PARTITION is issued in a
query. The first element of the list will be the parent table of that partition,
and next elements will be the CREATE/ALTER TABLE queries for partitions.

If this list does not have any statements other than the first one, or lc is
NULL (there is no statements to CREATE/ALTER partitions), further logic is
useless, since statements are needed to correctly process utility statements for
the partitions.