Truncate all segment files for temporary relations
It is an addition to commit 87dae12f1e9bfbf3c03b28961779243e826c59f0

File descriptors of segment files of temporary relation can be opened by
bg writer process. So, after unlink, space may not be released to the OS.
To fix it, truncate all segment files for all relations in mdunlinkfork,
including temporary.

Plus, the test case for temporary tables in the tablespace test was improved.
Amount of inserted data was reduced.