Assert that prule and pid2 must exist
If bExistError and bMustExist parameters are true, get_part_rule() throws an
error if partition rule does not exist.

For AT_AP_IDList kind:
- src/backend/cdb/cdbpartition.c:7600, get_part_rule() that will throw an error
  if partition does not exist
- src/backend/commands/tablecmds.c:16326, get_part_rule() that will throw an
  error if partition does not exist

For AT_AP_IDRule kind:
- src/backend/commands/tablecmds.c:16345, get_part_rule() that will throw an
  error if partition does not exist
- src/backend/commands/tablecmds.c:16050, asserting that both list entries exist

Assert that prule2 is not null.
Assert that pid are not null, so get_part_rule() can't return NULL with
bMustExist = true.
Assert that if get_part_rule() is used to get the partition rule, it is called
with bExistError = bMustExist = true.