Assert that prule and pid2 must exist
If bExistError and bMustExist parameters are true, get_part_rule1() throws an
error if partition rule does not exist.

For AT_AP_IDList kind, get_part_rule1() is used that will throw an error
if partition does not exist.

For AT_AP_IDRule kind:
- get_part_rule1() that will throw an error if partition does not exist.
- In wack_pid_relname() there is an assert that assumes both list
  entries exist.

In these cases get_part_rule1() is called with bExistError = bMustExist =
true, so it can't return NULL.

Assert that prule2 is not NULL.