add check for "parentPolicy is not NULL" before using
Pointer to "parentPolicy" is set from "parentrel->rd_cdbpolicy".
It can be NULL if it was not set at function "heap_create_with_catalog".
It depends on "policy" is NULL or not. NULL "policy" will not be copied to
"new_rel_desc->rd_cdbpolicy".

At other places at function "transformDistributedBy" this field is checked
before using. At the last "if" "parentPolicy" is used without checking. Missing
check was added.