Remove dependency to system calls for memory allocation in regress_gp
Failures in allocations could lead to crashes with NULL pointer
dereferences .  Memory context TopMemoryContext is used instead to keep
alive the plans allocated in the session.  A more specific context could
be used here, but this is left for later.

This is copy of commit 45e47b3, applied to regress_gp.