Add check for fopen function result
In some files, the debugging functions did not have a check for what the fopen
function returned, so I added such a check.

This patch also adds a warning when fopen fails, because these are dumping
helper functions and it wouldn't be very nice if they crashed on errors like
opening a file.