Add tableoid attribute to ORCA UPDATE/DELETE targetlist
ORCA's Split Update and Delete DML operations performed partition pruning for
each incoming tuple at delete stage. Executor always called slot_get_partition
to get the target partition to delete from. This behaviour did cause not only
the increased computational cost, but it could lead to data corruption when the
relation, from which tuple is extracted, did not correspond to relation defined
after partition pruning inside Update/Delete operations. For example, if target
row to delete was located in wrong partition (e.g. after ALTER TABLE EXCHANGE
PARTITION WITHOUT VALIDATION operation), then deletion will be issued using
ctid from that wrong partition on another right partition (which was chosen from
slot_get_partition). This scenario in worth case ends up to deletion of some
wrong row, i.e., data corruption.

This patch proposes solution of the issue described above. Instead of performing
partition selection at delete stage in ExecDelete, it would be more optimal to
add a tableoid attribite to the incoming slot. This attribute will contain the
oid of the leaf partition, which the tuple actually belongs to. In this case
there won't be any need to repeat the partiton pruning, because the tuple for
Update/Delete will already contain the oid of its relation. In order to achive
that the patch adds the tableoid atrribute to the ORCA's plan target lists for
Update/Delete plans. To add the new attribute to the final target list, this
patch adds a tableoid column reference both to DXL DML nodes (Query and Plan
DXL) and to each logical and physical DML operator. We must notice that tableoid
is added only for DML operations executed on root partitions (partitioned
tables). The corresponding Xforms and translators are also extended with the
tableoid handling. Because the new column have been added to the operator
classes, the parser classes (CParseHandler*) have been also extended in order
to support the proper print output for either debug print or minidumps.

On executor side the patch removes the partition pruning code from the
ExecDelete function. Is uses targetid_get_partition with tableoid attribute of
the slot, which is extracted in ExecDML function.

The tests have been added both to the regress and ORCA's minidump suites. Added
minidumps show the presence of tableoid attribute in DXL plan. Regress tests
check the correctnes of DML operations for modifying tuples belonging to wrong
partitions as it is described above.

This patch also modifies some code and tests from commit fd5939d in order to
adjust it to the tableoid presence.