Fix null check for ITask::Self
In several places, there is a check for ITask::Self() for null, but then CTask
is used. Self in CTask does dynamic_cast the result of ITask::Self(), which can
lead to null pointer dereference.

This patch fixes the null check by checking the result of CTask instead of
ITask, also this patch adds such a check in two other places.