Fix checks for partition rules (#702)
For AT_AP_IDList, add an assert that lfirst() result is not NULL.

For AT_AP_IDRule, return NULL if get_part_rule1() also returned NULL. If the
caller does not expect NULL to be returned, bExistError and bMustExist
parameters are set to true. In that case, get_part_rule1() can't return NULL and
will throw an error if rule does not exist before NULL is returned.