Fix potential NULL pointer dereferencing
In method CTranslatorExprToDXLUtils::PdxlnProjElem variable pdxlopPrElChild may
potentially be NULL, and it is used to invoke a method.

In CDXLNode m_dxl_op (returned by GetOperator and assigned to pdxlopPrElChild)
can be NULL only in one case - if the node was created with a 1-parameter ctor
CDXLNode(CMemoryPool *mp) and later CDXLNode::SetOperator was not called.

According to code, CDXLNode(CMemoryPool *mp) is called only from 2 places:
 - CParseHandlerScalarIdent::EndElement
 - CParseHandlerScalarConstValue::StartElement
In both places, ctor is immediately followed by CDXLNode::SetOperator, which
sets m_dxl_op to a not-NULL value (because it is checked in assert right before
CDXLNode creation). In all other CDXLNode ctors and CDXLNode::SetOperator the
operator passed as a parameter is checked to not be NULL.

As CDXLNode with NULL operator is not a valid runtime scenario and should never
happen under normal conditions, assert was added to the pdxlopPrElChild is not
NULL.