Fix the potential dereference of the null pointer in CErrorHandlerStandard
In the CErrorHandlerStandard method::Process is used by dynamic_cast to cast the
ILogger interface to the CLogger implementation. At the moment, the ILogger
interface has only one CLogger implementation, but in the future others may
appear and then dynamic_cast may return null and the null pointer will be
dereferenced.

This patch adds verification of the dynamic_cast result.