Fix possible segfault during parsing OptimizerConfig
When parsing OptimizerConfig from DXL, if the order of the elements in the
OptimizerConfig block does not match the expected order, we may end up with an
element of the wrong type, which will lead to a null pointer dereference.

This patch adds an assert to the dynamic_cast result.