Fix gprecoverseg behave test (#874)
In the scenario "gprecoverseg should not give warning if pg_basebackup is
running for the up segments", after stopping the primary segments, there is no
waiting for switching to mirrors, and therefore the step "the user suspend the
walsender on the primary on content 2" fails.

This patch adds the "user can start transactions" step after stop primary
segments to wait for switching to mirrors. As is done in other scenarios.

(cherry picked from commit e38f5722ab6ef83826ba95a25cfb4e4776ec96a6)