Filter excessive attrs from ORCA's DML delete node
Currently, ORCA passes all attributes to DML node regardless of the type of
operation. Passing all user-defined attrs to delete operation causes
unnecessary scans of these attrs. This may lead to performance decreasing,
especially for column-oriented tables. In the most of cases, such attrs may be
omitted. The only exception is the "after" trigger, which needs all attrs to
work.
The goal of this patch is to change the behavior of delete operation and use
empty source array for DML node if possible. To use it, execution part of DML
node was also changed. From now, it can work with subset of attrs in target
list, which is close to how postgres works.
We agreed that ORCA has enough tests for DELETE operation, including one with
triggers. But there is an interesting case with after trigger on table with
dropped column. The new mdp test file was added for this case. It's suitable for
ORCA unit tests, but a real execution of a script inside gives an error. This
bug is not related to our patch; separate ticket was created.