Fix the check for address availability when setting up UDP sockets in IPv6 disabled environments
Add `AI_ADDRCONFIG` flag to `getaddrinfo()` hints to force it to check for
availability of the address families. This flag will prevent `getaddrinfo()`
from returning addresses that belong to the IPv6 family when IPv6 is not
available on the host machine.

This problem is present only when `INTERCONNECT_ADDRESS_TYPE_WILDCARD` is used
as the interconnect address type. It does not affect 7.X, because the default
setting for interconnect address type was changed to
`INTERCONNECT_ADDRESS_TYPE_UNICAST`.