Fix resource group tests
Patch 52bbf95d91155517a711b193bbffbaec880e7389 increased memory consumption on
the coordinator and caused the resgroup_cpu_rate_limit test to fail.

This patch adds the ability to create a demo cluster on multiple hosts, which
allows you to create an environment close to the upstream.
Add a script to run resource group tests.