Make the behave tests more stable
When initializing containers for behave tests, ssh host keys are created, which
each container then scans using ssh-keyscan. Since all containers are
initialized in parallel, there is a race condition. If one of the containers has
not yet managed to create keys, while the other container has already started
the scanning stage, in this case, the container may not add the keys of all
containers to known_hosts, which will lead to the failure of the ssh connection
during the tests.

This patch takes out the key scanning stage separately from container
initialization, so that scanning occurs only after keys are created by all the
containers.