ao/co: fix an issue with COPY zero-column CO table (#1720)
The issue is that if the CO table once had column (and data), but later all
columns are dropped, then COPY this table would hang:

```
create table co(a int) using ao_column;
insert into co select * from generate_series(1,10);
alter table co drop column a;

copy co to stdout; -- this hangs forever
```

The is because aocs_getnext is not handling this case well: it would open the
segment files but skip the actually scanning (because no projected columns),
write no tupleslot values, and finally return TRUE which doesn't make any sense
as it did not read any tuple. So a loop to get next tuple would loop forever.
If we return FALSE here, we would stop the loop early.

Note that, however, this does not make COPY's behavior for CO to be the same as
heap or AO, where we would still copy out the same number of rows as in the
table, just no data in the row. In order to get it right, we can make a
targeted change to COPY TO so it leave the decision to the CO AM layer
regarding which column to project, which is also more consistent with SELECT *.

The change to aocs_getnext() would serve as just defensive checks.

Reviewed-by: Kevin Yeap <kyeap@vmware.com>
(cherry picked from commit 50b4438f9c28b52a29951d2be76a1a1ed6c38ca7)

Original commit was modified and adapted due to different sources of AOCS tables
at 7X and 6X. One more test was added to check case with dropped columns.
Note about aoco_beginscan_extractcolumns() was removed, because there is no
aoco_beginscan_extractcolumns() in 6X.