Fix segmentation fault with a lateral join
Problem description:
Some queries with lateral join cause segmentation fault. It happens if the query
has nested sub-query (with more than one level nesting) with groupby or limit,
and the lateral variable is accessed from it.

Cause:
Postgres planner for these queries creates a plan with slices, and lateral
variable is accessed from a different slice. Lateral variables can't be send
across motions. So, it causes segmentation fault.

Fix:
Force such queries to single query executor (by setting 'force_singleQE' in
PlannerConfig structure). Previously, this flag was set for queries with lateral
variable and direct sub-query with groupby or limit. This check was not perfect
as the direct sub-query could not contain access to the lateral variable, and it
could have its own sub-query with access to the lateral variable, which was not
checked. Fix is done by improving this check - now
'is_query_contain_limit_groupby' does recursive check for nested sub-queries and
checks that query with groupby or limit really accesses any lateral var.